requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.17
python-dotenv==1.0.1
httpx==0.27.2
itsdangerous==2.2.0
itsdangerous==2.2.0
fastapi
uvicorn[standard]
jinja2
python-multipart
httpx
python-dotenv
